chore: Remove bad symlink used for tests (#10527)
* chore: Remove bad symlink used for tests


* Mark CreateSymlink as helper function
