+++ reposerver/repository/repository_test.go
@@ -137,31 +137,6 @@ func newServiceWithCommitSHA(root, revision string) *Service {
 	return service
 }
 
-// createSymlink creates a symlink with name linkName to file destName in
-// workingDir
-func createSymlink(t *testing.T, workingDir, destName, linkName string) error {
-	oldWorkingDir, err := os.Getwd()
-	if err != nil {
-		return err
-	}
-	if workingDir != "" {
-		err = os.Chdir(workingDir)
-		if err != nil {
-			return err
-		}
-		defer func() {
-			if err := os.Chdir(oldWorkingDir); err != nil {
-				t.Fatal(err.Error())
-			}
-		}()
-	}
-	err = os.Symlink(destName, linkName)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
 func TestGenerateYamlManifestInDir(t *testing.T) {
 	service := newService("../../manifests/base")
 
@@ -2080,9 +2055,9 @@ func Test_getPotentiallyValidManifests(t *testing.T) {
 	t.Run("circular link should throw an error", func(t *testing.T) {
 		const testDir = "./testdata/circular-link"
 		require.DirExists(t, testDir)
-		require.NoError(t, createSymlink(t, testDir, "a.json", "b.json"))
+		require.NoError(t, fileutil.CreateSymlink(t, testDir, "a.json", "b.json"))
 		defer os.Remove(path.Join(testDir, "a.json"))
-		require.NoError(t, createSymlink(t, testDir, "b.json", "a.json"))
+		require.NoError(t, fileutil.CreateSymlink(t, testDir, "b.json", "a.json"))
 		defer os.Remove(path.Join(testDir, "b.json"))
 		manifests, err := getPotentiallyValidManifests(logCtx, "./testdata/circular-link", "./testdata/circular-link", false, "", "", resource.MustParse("0"))
 		assert.Empty(t, manifests)
@@ -2180,9 +2155,9 @@ func Test_findManifests(t *testing.T) {
 	t.Run("circular link should throw an error", func(t *testing.T) {
 		const testDir = "./testdata/circular-link"
 		require.DirExists(t, testDir)
-		require.NoError(t, createSymlink(t, testDir, "a.json", "b.json"))
+		require.NoError(t, fileutil.CreateSymlink(t, testDir, "a.json", "b.json"))
 		defer os.Remove(path.Join(testDir, "a.json"))
-		require.NoError(t, createSymlink(t, testDir, "b.json", "a.json"))
+		require.NoError(t, fileutil.CreateSymlink(t, testDir, "b.json", "a.json"))
 		defer os.Remove(path.Join(testDir, "b.json"))
 		manifests, err := findManifests(logCtx, "./testdata/circular-link", "./testdata/circular-link", nil, noRecurse, nil, resource.MustParse("0"))
 		assert.Empty(t, manifests)
