+++ test/fixture/path/files.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"testing"
 )
 
 // CopyDir copies the contents of a directory from 'src' to 'dest'
@@ -74,5 +75,31 @@ func copySingleFile(src string, dest string, mode os.FileInfo) error {
 		return err
 	}
 
+	return nil
+}
+
+// CreateSymlink creates a symlink with name linkName to file destName in
+// workingDir
+func CreateSymlink(t *testing.T, workingDir, destName, linkName string) error {
+	t.Helper()
+	oldWorkingDir, err := os.Getwd()
+	if err != nil {
+		return err
+	}
+	if workingDir != "" {
+		err = os.Chdir(workingDir)
+		if err != nil {
+			return err
+		}
+		defer func() {
+			if err := os.Chdir(oldWorkingDir); err != nil {
+				t.Fatal(err.Error())
+			}
+		}()
+	}
+	err = os.Symlink(destName, linkName)
+	if err != nil {
+		return err
+	}
 	return nil
 }
