--- util/app/path/testdata/abslink/abslink
@@ -1,1 +1,0 @@
-/somethingbad
